# Conflicts:
#	game-core/game-core-common/src/main/java/com/fulu/game/common/enums/PushMsgJumpTypeEnum.java
#	game-core/game-core-service-impl/src/main/java/com/fulu/game/core/service/impl/PushMsgServiceImpl.java

+++ game-app/game-app-controller/src/main/java/com/fulu/game/app/controller/DynamicController.java
@@ -65,7 +65,7 @@ public class DynamicController extends BaseController {
      * @return
      */
     @RequestMapping(value = "delete")
-    public Result delete(@RequestParam(required = true) Long id) {
+    public Result delete(@RequestParam(required = true) Integer id) {
         dynamicService.deleteDynamicById(id, true);
         return Result.success().msg("删除成功！");
     }
@@ -112,7 +112,7 @@ public class DynamicController extends BaseController {
      * @return
      */
     @RequestMapping(value = "detail")
-    public Result detail(@RequestParam( required = true) Long id) {
+    public Result detail(@RequestParam( required = true) Integer id) {
         DynamicDoc dynamicDoc = dynamicService.getDynamicDocById(id);
         return Result.success().data(dynamicDoc);
     }
