+++ game-core/game-core-entity/src/main/java/com/fulu/game/core/entity/DynamicComment.java
@@ -18,17 +18,17 @@ public class DynamicComment implements Serializable {
 	private static final long serialVersionUID = 1L;
 	
 	//id
-	private Long id;
+	private Integer id;
 	//动态id(用户查询评论的所有回复信息)
-	private Long dynamicId;
+	private Integer dynamicId;
 	//目标id（当reply_type=1时为评论id，回复的父id）
-	private Long commentId;
+	private Integer commentId;
 	//类型（1：评论动态；2：评论动态的评论）
 	private Integer commentType;
 	//评论的内容
 	private String content;
 	//用户id
-	private Long fromUserId;
+	private Integer fromUserId;
 	//用户头像URL（冗余字段，提高查询效率）
 	private String fromUserHeadUrl;
 	//用户昵称（冗余字段，提高查询效率）
@@ -36,7 +36,7 @@ public class DynamicComment implements Serializable {
 	//用户性别(默认0：不公开；1：男；2：女)
 	private Integer fromUserGender;
 	//目标用户id
-	private Long toUserId;
+	private Integer toUserId;
 	//目标用户头像URL（冗余字段，提高查询效率）
 	private String toUserHeadUrl;
 	//目标用户昵称（冗余字段，提高查询效率）
